feat/test: updated player points calculations to use BigDecimal To remove accuracy issues, and added test to check a simple case

+++ backend/src/main/kotlin/com/nicolaspetras/splfantasy/model/scoring/internal/SplPlayerMatchScore.kt
@@ -4,21 +4,22 @@ import com.nicolaspetras.splfantasy.model.SmiteRole
 import com.nicolaspetras.splfantasy.model.SplTeamName
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import java.math.BigDecimal
 
 class SplPlayerMatchScore(
     // TODO: Change to SplPlayer -- remove name, role and team
     val name: String = "",
     val role: SmiteRole = SmiteRole.NONE,
     val team: SplTeamName = SplTeamName.NONE,
-    var gameScores: ArrayList<Double> = arrayListOf()
+    var gameScores: ArrayList<BigDecimal> = arrayListOf()
 ) {
     private val log: Logger = LoggerFactory.getLogger(javaClass)
 
     /**
      * Returns the overall match score for the player
      */
-    fun overallMatchScore(): Double {
-        var cumulativeScore = 0.0
+    fun overallMatchScore(): BigDecimal {
+        var cumulativeScore = BigDecimal(0.0)
         if (gameScores.isNotEmpty()) {
             for (score in gameScores) {
                 cumulativeScore += score
